Add files via upload
DeskPins can be used to make any application topmost, that is, to keep it above all other windows. Just grab a pin from the DeskPins icon in the system notification area and click on any window. This functionality is identical to the "Always on Top" setting found in some programs. With DeskPins you can add this feature to any program.